requirements.txt
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.3
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.7
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
whitenoise==4.1
